Updating to reflect 2023 better
This page used to talk more about the Overall Rating for a performance review back when the overall rating was more directly tied to the Maturing/Performing/Exceeding names we use for bands.

We've since disconnected the overall rating for performance reviews, using a different set of names and values, so that Levels and Bands can be unambiguously their own thing.

I've also tried to push readers to the handbook and focus this page on calibrating what we mean by Maturing, Performing, Exceeding.