Clarifies options when struggling to reach Performing
Once someone is working towards a certain Level, ie; Maturing, it's important that the manager and team member are aligned on the trajectory toward Performing in that level. Misalignment tends to result in frustration for the manager and team member.

Imagine a scenario where the manager and team member are aligned, but the team member is still struggling to reach Performing. There are typically two key options available, namely a mutually agreeable support plan or a mutually agreed return to Exceeding at the previous level.

The important change to this wording is to be clear on the mutual agreement in either case.